Adapt RuneLite model loader
This was no doubt the original intent here. This requires us to use a
newer version of Gradle, as the older one isn't compatible with newer
javac. Maybe it can be ported to kotlin too, but there isn't much point
when it won't gain much readability in the first place.

Also fix a couple of caching-related bugs.