Simplify JVM args and use sh for ./run
This removes -XX:+SuppressFatalErrorMessage, which is a flag that
prevents the hs_err_pid crash dump from being created when a fatal error
occurs. It is an unnecessary break from defaults IMO.

This also removes the verbosity flag for JavaFX, because the information
is presents to us is spammy and not especially useful.

Otherwise some changes have been made to the script itself, namely:
- Invoking under sh instead of bash (more widely available, especially
  under /bin)
- Using `exec` to remove the shell middle-man once the java program is
  launched
- Using "$@" to pass arguments from the command line to the java
  program, not that we use those