Fix chunk stacking when renderer is dead
This removes the x/y/z info from a SceneTile, since that's relative to
a scene and it's more expensive to update it in the tile than it is to
just pass it around.

The issue was ultimately caused by the x and y coordinates of a
SceneTile being mutable, and not being set at a time when they were
expected to be.

Fix #104